Limit all persist_sink activators
